created adopt view, changins navigate in animals to /adopt
