+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Routes, Route, BrowserRouter } from "react-router-dom";
-import AdoptForm from "./views/AdoptForm";
+import AdoptForm from "./components/AdoptForm";
 import DonationList from "./views/DonationList";
 import Navbar from "./components/Navbar";
 import VolunteerForm from "./views/VolunteerForm";
@@ -12,7 +12,7 @@ import RefugeeRoutes from "./views/dashboard/refugee/RefugeeRoutes";
 import Login from "./components/Login";
 import Footer from "./components/Footer";
 import NotFound from "./components/notFound";
-import Adopt from "./views/adopt";
+import Adopt from "./views/Adopt";
 import Logout from "./components/Logout";
 import ForgotPassword from "./components/ForgotPassword";
 import RegisterForm from "./components/RegisterForm";
@@ -21,6 +21,7 @@ import RegisterPage from "./views/RegisterPage";
 import About from "./views/About";
 import PrivacyPolicy from "./views/PrivacyPolicy";
 import NotificationContainer from "./components/NotificationContainer";
+import AdoptPage from "./views/Adopt";
 
 function App() {
   return (
@@ -35,7 +36,7 @@ function App() {
           <Route path="/signin" element={<RegisterPage />} />
           <Route path="/donationlist" element={<DonationList />} />
           <Route path="/volunteerform" element={<VolunteerForm />} />
-          <Route path="/adoptform/:animal_id" element={<AdoptForm />} />
+          <Route path="/adopt/:animal_id" element={<AdoptPage />} />
           <Route path="/animalprofile/:id" element={<AnimalProfile />} />
           <Route
             path="/forgotpassword"
