+++ src/components/AnimalProfile.tsx
@@ -101,13 +101,12 @@ const AnimalProfile: React.FC = () => {
             Estado de adopción:
           </p>
           <p
-            className={`font-semibold text-sm sm:text-base px-4 py-2 rounded-full shadow-md ${
-              animal.adoption_status === "adoptado"
+            className={`font-semibold text-sm sm:text-base px-4 py-2 rounded-full shadow-md ${animal.adoption_status === "adoptado"
                 ? "bg-white text-teal-500 border-2 border-teal-500"
                 : animal.adoption_status === "disponible"
-                ? "bg-teal-500 text-white border-none"
-                : "bg-yellow-400 text-white border-none"
-            }`}
+                  ? "bg-teal-500 text-white border-none"
+                  : "bg-yellow-400 text-white border-none"
+              }`}
           >
             {animal.adoption_status}
           </p>
@@ -117,7 +116,7 @@ const AnimalProfile: React.FC = () => {
         <div className="flex flex-col sm:flex-row mt-8 justify-between items-center gap-4">
           {/* Botón formulario */}
           <button className="bg-teal-500 text-white py-3 px-6 rounded-lg shadow-md border-2 border-teal-500 text-lg tracking-widest cursor-pointer hover:bg-white hover:text-teal-500 hover:border-teal-500 hover:shadow-lg active:bg-[#87dbd0] transition duration-400 w-full sm:w-auto">
-            <Link to={`/adoptform/${id}`} className="flex items-center justify-center">
+            <Link to={`/adopt/${id}`} className="flex items-center justify-center">
               <span>Rellena el formulario</span>
             </Link>
           </button>
